+++ bitmovin/account/account.js
@@ -2,9 +2,9 @@ import urljoin from 'url-join';
 
 import http from '../utils/http';
 
-import organizations from './organizations/organizations';
-import billing from './billing/billing';
-import apiKeys from './apiKeys/apiKeys';
+import organizations from './organizations';
+import billing from './billing';
+import apiKeys from './apiKeys';
 
 export const account = (configuration, http) => {
   const {get, post} = http;
